+++ django/applications/catmaid/static/js/WindowMaker.js
@@ -930,6 +930,7 @@ var WindowMaker = new function()
     var connectorRestrictionsSl = document.createElement('select');
     connectorRestrictionsSl.options.add(new Option('All connectors', 'none', true, true));
     connectorRestrictionsSl.options.add(new Option('All shared connectos', 'all-shared'));
+    connectorRestrictionsSl.options.add(new Option('All pre->post connectos', 'all-pre-post'));
     connectorRestrictionsSl.onchange = function () {
       WA.setConnectorRestriction(this.value);
     };