+++ CHANGELOG.md
@@ -85,6 +85,16 @@ Tile layer:
 - New filter (WebGL only): intensity thresholded transparency
 
 
+3D Viewer:
+
+- Connector restriction can now explicitly be turned on and off with a pull
+  down list. One can select between "Show all" (i.e. turned restriction turned
+  off), "All shared connectors" will only show connectors with partners in the
+  current selection and "All pre->post connectors" will only allow connectors
+  with at least one presynaptic partner and one postsynaptic partner in the
+  current selection.
+
+
 Neuron search:
 
 - The name displayed for neurons now follows the same naming mechanism used for